+++ pgvector/peewee/__init__.py
@@ -1,4 +1,14 @@
-from ..utils import SparseVec
+from .bit import FixedBitField
 from .halfvec import HalfvecField
 from .sparsevec import SparsevecField
 from .vector import VectorField
+from ..utils import HalfVec, SparseVec
+
+__all__ = [
+    'VectorField',
+    'HalfvecField',
+    'FixedBitField',
+    'SparsevecField',
+    'HalfVec',
+    'SparseVec'
+]